requirements.txt
plumbum>=1.6.9
numpy>=1.20.0
pandas>=1.2.1
seaborn>=0.11.1
matplotlib>=3.3.4
altair>=4.1.0
streamlit>=0.75.0
dash==1.19.0
jupyter==1.0.0
biopython==1.78
pylint==2.6.0